requirements.txt
#These are taken from a different project I was working on.

asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
psycopg==3.2.4
psycopg-binary==3.2.4
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.9.0
